ai_chat/requirements.txt
fastapi[standard]==0.115.0
uvicorn==0.30.6
google-generativeai==0.8.2
faiss-cpu==1.8.0.post1
numpy==1.26.4
pypdf==5.0.1
python-docx==1.1.2
python-multipart==0.0.12
